HDDS-11576. Create a separate S3 client factory
